+++ lib/src/base_hero_card.dart
@@ -1,44 +1,54 @@
-/*File : local_hero_transform
-Version : 0.0.1
+/* File: local_hero_transform
+   Version: 0.0.2
 */
+
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 import 'hero_card_model.dart';
 
+/// BaseFavoriteCard widget that displays a single card with various details and animations.
 class BaseFavoriteCard extends StatelessWidget {
-  final BaseHeroCardParameters parameters;
+  final BaseHeroCardParameters parameters; // Parameters required to build the card.
+
+  /// Constructor for BaseFavoriteCard, requiring the parameters.
   const BaseFavoriteCard({super.key, required this.parameters});
 
   @override
   Widget build(BuildContext context) {
+    // Determine the text direction for layout.
     bool isRtl = parameters.textDirection == TextDirection.rtl;
-    BaseHeroCardOptionalParameters? optionalParams = parameters.optionalParameters;
+    BaseHeroCardOptionalParameters? optionalParams =
+        parameters.optionalParameters; // Optional parameters for customization.
+
     return ConstrainedBox(
-      constraints: BoxConstraints(maxHeight: parameters.heightImage),
+      constraints:
+          BoxConstraints(maxHeight: parameters.heightImage), // Constrain the card's maximum height.
       child: Container(
         decoration: BoxDecoration(
           border: Border.all(
-            color: Colors.white.withValues(alpha: 0.1),
+            color: Colors.white.withValues(alpha: 0.1), // Light border around the card.
           ),
           boxShadow: [
             BoxShadow(
               blurRadius: 10,
-              color: Color(0xFF3A5160).withValues(alpha: 0.1),
-              offset: const Offset(0, 5),
+              color: Color(0xFF3A5160).withValues(alpha: 0.1), // Shadow effect for depth.
+              offset: const Offset(0, 5), // Position of the shadow.
             ),
           ],
-          color: Colors.white,
-          borderRadius: BorderRadius.circular(24.0.r),
+          color: Colors.white, // Background color of the card.
+          borderRadius: BorderRadius.circular(24.0.r), // Rounded corners.
         ),
-        margin: EdgeInsets.symmetric(horizontal: 10.w, vertical: 8.h),
+        margin: EdgeInsets.symmetric(horizontal: 10.w, vertical: 8.h), // Margin for spacing.
         child: Stack(
           children: [
+            // Display the image or a custom image widget if provided.
             optionalParams?.imageWidget ??
                 _buildImage(
                   height: parameters.heightImage,
                   width: parameters.widthImage,
                 ),
+            // Position the favorite icon button on the card.
             Positioned(
               top: parameters.favoriteIconHeightPosition.abs(),
               right: isRtl ? parameters.favoriteIconPosition : null,
@@ -46,18 +56,19 @@ class BaseFavoriteCard extends StatelessWidget {
               child: optionalParams?.favoriteIconButton ??
                   IconButton(
                     style: ButtonStyle(
-                      backgroundColor: WidgetStatePropertyAll(
-                        Colors.redAccent,
-                      ),
+                      backgroundColor:
+                          WidgetStateProperty.all(Colors.redAccent), // Background for the button.
                     ),
                     icon: Icon(
                       Icons.favorite,
                       color: Colors.white,
                       size: 24,
                     ),
-                    onPressed: optionalParams?.onPressedFavoriteIcon,
+                    onPressed:
+                        optionalParams?.onPressedFavoriteIcon, // Action when the button is pressed.
                   ),
             ),
+            // Display the card name.
             Positioned(
               bottom: parameters.bottomTitle,
               right: isRtl ? parameters.rightTitle : 10.w,
@@ -73,10 +84,11 @@ class BaseFavoriteCard extends StatelessWidget {
                           fontWeight: FontWeight.w500,
                         ),
                       ),
-                      const Spacer(),
+                      const Spacer(), // Spacer to push content to the left.
                     ],
                   ),
             ),
+            // Display the card title.
             Positioned(
               bottom: (parameters.bottomTitle / 1.6).h,
               right: isRtl ? parameters.rightPrice : null,
@@ -91,9 +103,9 @@ class BaseFavoriteCard extends StatelessWidget {
                         fontWeight: FontWeight.w500,
                       ),
                       children: [
-                        WidgetSpan(child: SizedBox(width: 3.w)),
+                        WidgetSpan(child: SizedBox(width: 3.w)), // Space between text elements.
                         TextSpan(
-                          text: 'sar',
+                          text: 'sar', // Currency symbol.
                           style: TextStyle(
                             color: Colors.black,
                             fontSize: 14.sp,
@@ -104,11 +116,13 @@ class BaseFavoriteCard extends StatelessWidget {
                     ),
                   ),
             ),
+            // Display the subtitle of the card.
             Positioned(
               bottom: parameters.bottomTitle / 6,
               right: isRtl ? parameters.rightPrice : null,
               left: !isRtl ? parameters.rightPrice : null,
-              child: optionalParams?.subtitleWidget ?? _buildSubTitle(context),
+              child: optionalParams?.subtitleWidget ??
+                  _buildSubTitle(context), // Use a subtitle widget if provided.
             ),
           ],
         ),
@@ -116,57 +130,63 @@ class BaseFavoriteCard extends StatelessWidget {
     );
   }
 
+  /// Method to build the card image.
   Widget _buildImage({double? height, double? width}) {
     return Padding(
-      padding: EdgeInsets.all(8.0.h),
+      padding: EdgeInsets.all(8.0.h), // Padding around the image.
       child: ClipRRect(
-        borderRadius: BorderRadius.circular(16.r),
+        borderRadius: BorderRadius.circular(16.r), // Rounded corners for the image.
         child: Image.network(
-          parameters.cardModel.imageUrl,
-          height: height?.h ?? 200,
-          width: width?.w ?? 200,
-          fit: BoxFit.cover,
+          parameters.cardModel.imageUrl, // Image URL from the card model.
+          height: height?.h ?? 200, // Height of the image.
+          width: width?.w ?? 200, // Width of the image.
+          fit: BoxFit.cover, // Cover the entire area.
         ),
       ),
     );
   }
 
+  /// Method to build the subtitle for the card.
   Row _buildSubTitle(BuildContext context) {
     return Row(
-      mainAxisAlignment: MainAxisAlignment.start,
+      mainAxisAlignment: MainAxisAlignment.start, // Align items to the start.
       children: [
         Icon(
           Icons.location_on_outlined,
-          color: const Color(0xFF95979A),
+          color: const Color(0xFF95979A), // Color for the location icon.
           size: 10,
         ),
-        SizedBox(width: 1.w),
+        SizedBox(width: 1.w), // Space between icon and text.
         Text(
-          parameters.cardModel.name,
+          parameters.cardModel.name, // Display the card name as subtitle.
           style: TextStyle(
             color: const Color(0xFF95979A),
             fontSize: 13.sp,
             fontWeight: FontWeight.w400,
           ),
-          overflow: TextOverflow.ellipsis,
+          overflow: TextOverflow.ellipsis, // Handle overflow with ellipsis.
         ),
       ],
     );
   }
 }
 
+/// Parameters required to build a BaseFavoriteCard.
 class BaseHeroCardParameters {
-  final int index;
-  final double heightImage;
-  final double widthImage;
-  final double bottomTitle;
-  final double rightTitle;
-  final double rightPrice;
-  final double favoriteIconPosition;
-  final double favoriteIconHeightPosition;
-  final HeroCardModel cardModel;
-  final TextDirection textDirection;
-  final BaseHeroCardOptionalParameters? optionalParameters;
+  final int index; // Index of the card.
+  final double heightImage; // Height of the card image.
+  final double widthImage; // Width of the card image.
+  final double bottomTitle; // Position of the bottom title.
+  final double rightTitle; // Position of the right title.
+  final double rightPrice; // Position of the right price.
+  final double favoriteIconPosition; // Position of the favorite icon.
+  final double favoriteIconHeightPosition; // Height position of the favorite icon.
+  final HeroCardModel cardModel; // Model representing the card data.
+  final TextDirection textDirection; // Text direction for the card layout.
+  final BaseHeroCardOptionalParameters?
+      optionalParameters; // Optional parameters for customization.
+
+  /// Constructor for BaseHeroCardParameters, requiring necessary parameters.
   const BaseHeroCardParameters({
     this.optionalParameters,
     required this.index,
@@ -182,14 +202,16 @@ class BaseHeroCardParameters {
   });
 }
 
+/// Optional parameters for customizing the BaseFavoriteCard.
 class BaseHeroCardOptionalParameters {
-  final Widget? titleWidget;
-  final Widget? nameWidget;
-  final Widget? favoriteIconButton;
-  final VoidCallback? onPressedFavoriteIcon;
-  final Widget? subtitleWidget;
-  final Widget? imageWidget;
+  final Widget? titleWidget; // Custom title widget.
+  final Widget? nameWidget; // Custom name widget.
+  final Widget? favoriteIconButton; // Custom favorite icon button.
+  final VoidCallback? onPressedFavoriteIcon; // Callback for favorite icon press.
+  final Widget? subtitleWidget; // Custom subtitle widget.
+  final Widget? imageWidget; // Custom image widget.
 
+  /// Constructor for BaseHeroCardOptionalParameters.
   BaseHeroCardOptionalParameters({
     this.titleWidget,
     this.nameWidget,