+++ lib/src/card_grid_view.dart
@@ -1,15 +1,26 @@
-/*File : local_hero_transform
-Version : 0.0.1
+/* File: local_hero_transform
+   Version: 0.0.2
 */
+
 import 'package:flutter/material.dart';
 
 import '../local_hero_transform.dart';
 
+/// CardGridView widget that displays a card in a grid layout with hero animations.
 class CardGridView extends StatelessWidget {
+  /// Unique tag for the hero animation based on the index.
   final int index;
+
+  /// Model representing the data for the card.
   final HeroCardModel cardModel;
+
+  /// Direction displayed the card.
   final TextDirection textDirection;
+
+  /// Optional parameters for additional customization.
   final BaseHeroCardOptionalParameters? optionalParameters;
+
+  // Constructor for CardGridView, requiring card model, index, text direction, and optional parameters.
   const CardGridView({
     required this.cardModel,
     required this.index,
@@ -24,17 +35,21 @@ class CardGridView extends StatelessWidget {
       tag: index,
       flightShuttleBuilder:
           (flightContext, animation, flightDirection, fromHeroContext, toHeroContext) {
+        // Define animations for the position of the card during the transition.
         final positionRight = Tween<double>(begin: 100, end: 10).animate(animation);
         final positionBottom = Tween<double>(begin: 50, end: 50).animate(animation);
 
-        // Get size for the to hero widget or from hero widget size
+        // Get sizes of the hero widgets from their render boxes.
         RenderBox? renderBoxFrom = fromHeroContext.findRenderObject() as RenderBox?;
         RenderBox? renderBoxTo = toHeroContext.findRenderObject() as RenderBox?;
+
+        // Define animations for height and width of the card during transition.
         final animationHeight =
             Tween<double>(begin: 90, end: renderBoxTo!.size.height * 0.67).animate(animation);
-
         final animationWidth =
             Tween<double>(begin: 80, end: renderBoxTo.size.width * 0.9).animate(animation);
+
+        // Define animations for the favorite icon's position.
         final favoriteIconPosition =
             Tween<double>(begin: renderBoxFrom!.size.width - 72, end: 20).animate(animation);
         final favoriteIconHeightPosition =
@@ -43,42 +58,49 @@ class CardGridView extends StatelessWidget {
         return AnimatedBuilder(
           animation: animation,
           builder: (context, child) {
+            // Build the animated card using BaseFavoriteCard with updated parameters.
             return BaseFavoriteCard(
               parameters: BaseHeroCardParameters(
                 optionalParameters: optionalParameters,
                 textDirection: textDirection,
-                favoriteIconHeightPosition: favoriteIconHeightPosition.value,
-                favoriteIconPosition: favoriteIconPosition.value,
-                cardModel: cardModel,
-                index: index,
-                heightImage: animationHeight.value,
-                widthImage: animationWidth.value,
-                bottomTitle: positionBottom.value,
-                rightTitle: positionRight.value,
-                rightPrice: positionRight.value,
+                favoriteIconHeightPosition: favoriteIconHeightPosition
+                    .value, // Animated position for the favorite icon's height.
+                favoriteIconPosition:
+                    favoriteIconPosition.value, // Animated position for the favorite icon.
+                cardModel: cardModel, // Card model data.
+                index: index, // Index of the card.
+                heightImage: animationHeight.value, // Animated height of the card image.
+                widthImage: animationWidth.value, // Animated width of the card image.
+                bottomTitle: positionBottom.value, // Animated bottom position of the title.
+                rightTitle: positionRight.value, // Animated right position of the title.
+                rightPrice: positionRight.value, // Animated right position of the price.
               ),
             );
           },
         );
       },
+      // Create a rectangular tween for the hero animation.
       createRectTween: (Rect? begin, Rect? end) {
-        return MaterialRectCenterArcTween(begin: begin, end: end);
+        return MaterialRectCenterArcTween(
+            begin: begin, end: end); // Arc tween for smooth transitions.
       },
       child: LayoutBuilder(
         builder: (context, constraints) {
+          // Build the base card that is displayed initially.
           return BaseFavoriteCard(
             parameters: BaseHeroCardParameters(
               textDirection: textDirection,
-              favoriteIconHeightPosition: 20,
-              favoriteIconPosition: 20,
-              cardModel: cardModel,
-              optionalParameters: optionalParameters,
-              index: index,
-              heightImage: constraints.maxHeight * 0.67,
-              widthImage: 180,
-              bottomTitle: 50,
-              rightTitle: 10,
-              rightPrice: 10,
+              favoriteIconHeightPosition: 20, // Initial height for the favorite icon position.
+              favoriteIconPosition: 20, // Initial position for the favorite icon.
+              cardModel: cardModel, // Card model data.
+              optionalParameters: optionalParameters, // Optional parameters.
+              index: index, // Index of the card.
+              heightImage: constraints.maxHeight *
+                  0.67, // Set height for the card based on layout constraints.
+              widthImage: 180, // Fixed width for the card image.
+              bottomTitle: 50, // Bottom position for the title.
+              rightTitle: 10, // Right position for the title.
+              rightPrice: 10, // Right position for the price.
             ),
           );
         },